// web/src/api/generated/models/StoryOrder.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const StoryOrder = {
    ASC: 'ASC',
    DESC: 'DESC'
} as const;
export type StoryOrder = typeof StoryOrder[keyof typeof StoryOrder];


export function instanceOfStoryOrder(value: any): boolean {
    for (const key in StoryOrder) {
        if (Object.prototype.hasOwnProperty.call(StoryOrder, key)) {
            if (StoryOrder[key as keyof typeof StoryOrder] === value) {
                return true;
            }
        }
    }
    return false;
}

export function StoryOrderFromJSON(json: any): StoryOrder {
    return StoryOrderFromJSONTyped(json, false);
}

export function StoryOrderFromJSONTyped(json: any, ignoreDiscriminator: boolean): StoryOrder {
    return json as StoryOrder;
}

export function StoryOrderToJSON(value?: StoryOrder | null): any {
    return value as any;
}

export function StoryOrderToJSONTyped(value: any, ignoreDiscriminator: boolean): StoryOrder {
    return value as StoryOrder;
}